# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Organization service for create/update/delete organization for user. It allows to invite/remove other users, assign them role and update their roles.",
        "title": "Organization-Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000/",
    "paths": {
...
        "/organization/": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "update organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization"
                ],
                "summary": "update organization",
                "operationId": "update-organization",
                "parameters": [
                    {
                        "description": "The input for update organization",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateOrganizationDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "create organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization"
                ],
                "summary": "create organization",
                "operationId": "create-organization",
                "parameters": [
                    {
                        "description": "The input for create organization",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateOrganization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OrganizationID"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        },
        "/organization/delete/otp": {
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "otp for delete organization on owners's registered mail/phoneNumber",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organization"
                ],
                "summary": "otp for delete organization",
                "operationId": "delete-organization",
                "parameters": [
                    {
                        "description": "The input for update organization",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OrganizationID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.CustomError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error.CustomError": {
            "type": "object",
            "required": [
                "errorMessage",
                "statusCode"
            ],
            "properties": {
                "errorMessage": {
                    "type": "string"
                },
                "invalidData": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/error.InvalidData"
                    }
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "error.InvalidData": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "field": {
                    "type": "string"
                }
            }
        },
        "request.CreateOrganization": {
            "type": "object",
            "properties": {
                "Privacy": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "request.OrganizationID": {
            "type": "object",
            "required": [
                "organizationID"
            ],
            "properties": {
                "organizationID": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateOrganizationDetails": {
            "type": "object",
            "required": [
                "organizationID"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "organizationID": {
                    "type": "string"
                },
                "privacy": {
                    "type": "string"
                }
            }
        },
...
        "response.OrganizationID": {
            "type": "object",
            "properties": {
                "organizationID": {
                    "type": "string"
                }
            }
        },
...
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "UserIDAuth": {
            "type": "apiKey",
            "name": "Auth-user",
            "in": "header"
        },
        "jwtAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}